+++ database/internal/rdbms/orm/gorm/gorm.go
@@ -3,7 +3,7 @@ package gorm
 import (
 	"context"
 
-	"github.com/shashank-priyadarshi/utilities/database/models"
+	"go.ssnk.in/utils/database/models"
 	"gorm.io/gorm"
 )