+++ database/internal/redis/redis.go
@@ -4,10 +4,8 @@ import (
 	"context"
 	"time"
 
-	"github.com/shashank-priyadarshi/utilities"
-
 	"github.com/redis/go-redis/v9"
-	"github.com/shashank-priyadarshi/utilities/database/models"
+	"go.ssnk.in/utils/database/models"
 )
 
 type Handler struct {