+++ network/ports/ports.go
@@ -1,19 +1,22 @@
 package ports
 
-import "github.com/shashank-priyadarshi/utilities/network/ports/application"
+import "io"
 
-// Network is the main interface that all network types should implement
 type Network interface {
-	Protocol
+	Dialector
+	Server
+	Closer
 }
 
-// Protocol is the interface that all protocols should implement
-type Protocol interface {
-	application.HTTP
+type Dialector interface {
+	Dial() (io.ReadCloser, error)
 }
 
-// Connection interface might be removed altogether
-//type Connection interface {
-//	Connect(...interface{}) error
-//	Disconnect(...interface{}) error
-//}
+type Server interface {
+	Configure() (interface{}, error)
+	Start() error
+}
+
+type Closer interface {
+	Close() error
+}