+++ test/integration/modules/network/network.go
@@ -1,12 +1,12 @@
 package network
 
 import (
-	"github.com/shashank-priyadarshi/utilities/network"
-	"github.com/shashank-priyadarshi/utilities/network/models"
+	"go.ssnk.in/utils/network"
+	"go.ssnk.in/utils/network/types"
 )
 
 func Test() {
-	netw, _ := network.New(&models.Config{})
+	netw, _ := network.New(&types.Config{})
 	client, _ := netw.Start()
 	group, _ := client.Group()
 	_ = group.Handler()