+++ test/load/load.go
@@ -1,7 +1,7 @@
 package load
 
 import (
-	"github.com/shashank-priyadarshi/utilities/test/types"
+	"go.ssnk.in/utils/test/types"
 )
 
 type Load struct{}